refactor(runner): move start args and run logic from main.rs to runner.rs
Keep main.rs as a thin CLI dispatch layer (~70 lines). RunConfig becomes
module-private in runner.rs. Also removes the unused --strict flag from
the setup subcommand.
